refactor(core)!: reduce the public API to an embedding facade
vuio-core is about to be published for device integrators who ship it inside
firmware, where a breaking change reaches installations slowly or never. The
crate exposed 356 publicly reachable items across 17 modules; the only consumer
in the tree, vuio-cli, used three of them. A surface that size cannot be kept
stable, so commit to a small one instead.

The public API is now:

    Runtime, RuntimeHandle, RuntimeOptions, RuntimeStatus
    Error, ErrorKind, Result

Everything else is pub(crate). Hosts needing richer interaction drive the
server over its HTTP and MCP APIs, which is already how vuio-tower works.

Specifics:

- One entry point. VuioRuntime and lifecycle::ApplicationRunner did the same
  job; Runtime remains and the CLI drives shutdown through RuntimeHandle rather
  than threading a CancellationToken in from outside.
- A real error type. anyhow::Error told callers nothing they could match on.
  Error carries a #[non_exhaustive] ErrorKind and erases its source to a boxed
  std::error::Error, so no dependency appears in the signature.
- No foreign types in the facade. RuntimeOptions no longer exposes
  tokio-util's CancellationToken — a 0.8 release of a transitive dependency
  would otherwise have been a breaking change for every integrator.
- RuntimeOptions is a builder with private fields, so new options are additive.
  Its three real overrides (port, server name, media directories) are folded
  into a config override internally, which keeps AppConfig and its 36 public
  fields out of the API entirely.

Integration tests reach into the DLNA, SSDP and database layers directly, so
an `unstable-internals` feature re-opens the modules for them. It carries no
stability promise and dependent crates must not enable it.

Making the modules internal revealed ~110 items with no caller inside the
crate — they existed only as API. They are kept compiling by a crate-level
allow(dead_code) in the default build; deleting them needs a per-item check on
every target platform, so it is left as follow-up rather than done blind on a
macOS build.
