+++ crates/vuio-core/src/lifecycle/media.rs
@@ -11,4 +11,3 @@ pub use events::ApplicationStats;
 pub(in crate::lifecycle) use events::*;
 pub(in crate::lifecycle) use monitoring::*;
 pub(in crate::lifecycle) use scanning::*;
-pub use service::MediaLifecycleService;
