+++ crates/vuio-core/src/web/mcp.rs
@@ -34,7 +34,7 @@ const MCP_CLIENT_TTL: Duration = Duration::from_secs(30 * 60);
 const MCP_MAX_RESPONSE_BYTES: usize = 1024 * 1024;
 
 mod protocol;
-pub use protocol::{JsonRpcError, JsonRpcRequest, JsonRpcResponse};
+pub use protocol::{JsonRpcRequest, JsonRpcResponse};
 
 // ──────────────────────────────────────────
 // JSON-RPC 2.0 types
@@ -50,7 +50,9 @@ mod transport;
 
 use catalog::*;
 use dispatch::*;
-pub use tools::{cached_renderers, cast_file_helper, cast_playlist_helper, cast_tracks_helper};
+pub use tools::{cast_file_helper, cast_playlist_helper, cast_tracks_helper};
+// MessageQuery is exercised by the MCP integration tests.
+#[allow(unused_imports)]
 pub use transport::{message_handler, sse_handler, MessageQuery};
 
 #[cfg(test)]
