+++ crates/vuio-core/src/web/mcp/tools.rs
@@ -3,6 +3,6 @@ mod media;
 mod playlists;
 
 pub(crate) use casting::*;
-pub use casting::{cached_renderers, cast_file_helper, cast_playlist_helper, cast_tracks_helper};
+pub use casting::{cast_file_helper, cast_playlist_helper, cast_tracks_helper};
 pub(crate) use media::*;
 pub(crate) use playlists::*;
